Переместите пременные
