+++ js/dom-util.js
@@ -8,8 +8,6 @@ const SIMILAR_CARD_TEMPLATE = document.querySelector('#card')
   .querySelector('.popup');
 
 const MAP_FILTERS = document.querySelector('.map__filters');
-const TITLE = AD_FORM.querySelector('#title');
-const PRICE = AD_FORM.querySelector('#price');
 
 const FORMS = [{
   element: AD_FORM,
@@ -75,8 +73,6 @@ const activeForms = () => toggleForms(true);
 export {
   MAP_CANVAS,
   SIMILAR_CARD_TEMPLATE,
-  TITLE,
-  PRICE,
   removeExtra,
   fillPhotoOrDelete,
   deactiveForms,
